+++ modules/contracts/src.sol/transferDefinitions/HashlockTransfer.sol
@@ -1,15 +1,15 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.7.1;
-pragma experimental "ABIEncoderV2";
+pragma experimental ABIEncoderV2;
 
-import "../interfaces/ITransferDefinition.sol";
+import "./TransferDefinition.sol";
 
 /// @title Hashlock Transfer
 /// @notice This contract allows users to claim a payment locked in
 ///         the application if they provide the correct preImage. The payment is
 ///         reverted if not unlocked by the timelock if one is provided.
 
-contract HashlockTransfer is ITransferDefinition {
+contract HashlockTransfer is TransferDefinition {
   struct TransferState {
     bytes32 lockHash;
     uint256 expiry; // If 0, then no timelock is enforced
@@ -19,21 +19,9 @@ contract HashlockTransfer is ITransferDefinition {
     bytes32 preImage;
   }
 
-  string StateEncoding = "tuple(bytes32 lockHash, uint256 expiry)";
-
-  string ResolverEncoding = "tuple(bytes32 preImage)";
-
-  string Name = "HashlockTransfer";
-
-  function getRegistryInformation() external override view returns (RegisteredTransfer memory) {
-    RegisteredTransfer memory info = RegisteredTransfer({
-      name: Name,
-      stateEncoding: StateEncoding,
-      resolverEncoding: ResolverEncoding,
-      definition: address(this)
-    });
-    return info;
-  }
+  string public constant override Name = "HashlockTransfer";
+  string public constant override StateEncoding = "tuple(bytes32 lockHash, uint256 expiry)";
+  string public constant override ResolverEncoding = "tuple(bytes32 preImage)";
 
   function create(bytes calldata encodedBalance, bytes calldata encodedState) external override view returns (bool) {
     TransferState memory state = abi.decode(encodedState, (TransferState));
@@ -41,7 +29,7 @@ contract HashlockTransfer is ITransferDefinition {
 
     require(balance.amount[1] == 0, "Cannot create hashlock transfer with nonzero recipient balance");
     require(state.lockHash != bytes32(0), "Cannot create hashlock transfer with empty lockHash");
-    require(state.expiry > block.number || state.expiry == 0, "Cannot create hashlock transfer with expired timelock");
+    require(state.expiry == 0 || state.expiry > block.number, "Cannot create hashlock transfer with expired timelock");
     return true;
   }
 
@@ -56,7 +44,7 @@ contract HashlockTransfer is ITransferDefinition {
 
     // If you pass in bytes32(0), payment is canceled
     // If timelock is nonzero and has expired, payment is canceled
-    if (resolver.preImage != bytes32(0) && (state.expiry > block.number || state.expiry == 0)) {
+    if (resolver.preImage != bytes32(0) && (state.expiry == 0 || state.expiry > block.number)) {
       // Check hash for normal payment unlock
       bytes32 generatedHash = sha256(abi.encode(resolver.preImage));
       require(state.lockHash == generatedHash, "Hash generated from preimage does not match hash in state");